ai-service/requirements.txt
fastapi
uvicorn
mysql-connector-python
sentence-transformers
scikit-learn
numpy